A starting to come together tags solution
The goal is to decouple specific database primary keys for categories
from the things the user actually has, and to do everything through
external id tags. In doing so, it should be easier for users to
re-assign their bottles to the categories which are most relevant.

The UI almost certainly doesn't work on this commit, but we're
starting to add tests and things seem to mostly work there