The massive cocktails form improvements
This vendors in the select2 library, which seems to be a relatively
ok multiselect that works. It depends on jquery which is a bit sad
but oh well. There are a couple cdn'd assets here that I'd prefer
not to have, but again, oh well. This also retrofits the cocktail
ingredient submission to be less tied between the view and the
controller, which is much nicer.