Move modal into shared lib, and set it up on other pages
There's a fair amount of ugliness and "barely working" to this,
since I'm registering a million event listeners that I'm sure
aren't being cleaned up property with turbolinks. Plus all
the business about using the CocktailsController for shared
cocktails. Plus the availability not refreshing when you use
too much of a bottle. Plus I bet submitting the modal breaks
for cocktails which you don't have enough of an ingredient.

Anyway, it works in the happy path so lets stop there for now