Somehow, upgrade to Rails 7
This is a tiny simple app, so of course this just works except
for somehow modals not closing when you call `dispose()`. That
breaking makes literally no sense here, but now `hide()` does
work so I guess we'll use that instead. Will I regret the
cavalier nature of this? Maybe. But also I'm getting ready
to switch to turbo so that's something