Merge the signed in cocktail experience to include all drinks
I've ended up wanting this as I've used the app. It's annoying
to have to remember if a drink is a shared drink or one of my
recipes when going to log a drink, so this puts it all in one
place. I did add some small JS with this change, which is "bad"
considering my desire to move to hotwire + stimulus, but it's
very small