Try out the administrate gem
I had been baking admin-like functionality right into the rest of
the app, but that's pretty limiting and slow since I'm the only
admin. This pure db view is easier, and this particular gem
seems the most straightforward since it does the least. `rails_admin`
has lots of features, but it clearly doesn't seem to work with
webpacker and I'm not ready to move off of that yet